pyproject.toml
[tool.poetry]
name = "repl_python3_DC-Bot"
version = "0.1.0"
description = ""
authors = ["QuinnTheCoder <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
python-dotenv = "^0.19.0"
youtube_dl = "^2021.6.6"
Flask = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
